ajuste nos slicers
