+++ shopping_list_front/src/store/reducer_dtos/AuthDto.ts
@@ -1,0 +1,3 @@
+export interface AuthDto {
+  sessionToken: string
+}
