+++ shopping_list_front/src/store/reducer_dtos/UserDto.ts
@@ -1,0 +1,7 @@
+export interface UserDto {
+  userId: string,
+  displayName: string | null,
+  email: string | null,
+  phoneNumber: string | null,
+  photoURL: string | null
+}
