+++ shopping_list_front/src/store/store.ts
@@ -1,9 +1,11 @@
 import { configureStore } from "@reduxjs/toolkit";
-import { authReducer } from "../auth/authSlice";
+import { authReducer } from "./store_slice/authSlice";
+import { userReducer } from "./store_slice/userSlice";
 
 export const store = configureStore({
   reducer: {
     auth: authReducer,
+    user: userReducer,
   }
 })
 
@@ -11,3 +13,4 @@ export const store = configureStore({
 export type RootState = ReturnType<typeof store.getState>
 // Inferred type: {posts: PostsState, comments: CommentsState, users: UsersState}
 export type AppDispatch = typeof store.dispatch
+
