+++ shopping_list_front/src/store/store_slice/authSlice.ts
@@ -1,23 +1,14 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
-import { RootState } from "../store/store";
-
-export interface AuthDto {
-  userId: string,
-  sessionToken: string,
-  displayName: string | null,
-  email: string | null,
-  phoneNumber: string | null,
-  photoURL: string | null
-}
+import { RootState } from "../store";
+import { AuthDto } from "../reducer_dtos/AuthDto";
 
 export const authSlice = createSlice({
   name: "auth",
   initialState: {
-    userId: '',
     sessionToken: '',
   } as AuthDto,
   reducers: {
-    setUser: (state, action: PayloadAction<AuthDto>) => {
+    setAuth: (state, action: PayloadAction<AuthDto>) => {
       (Object.keys(action.payload) as (keyof AuthDto)[]).forEach((key) => {
         if (!action.payload[key] || state[key] === action.payload[key]) return;
         state[key] = action.payload[key];
@@ -26,6 +17,6 @@ export const authSlice = createSlice({
   },
 });
 
-export const { setUser } = authSlice.actions;
+export const { setAuth } = authSlice.actions;
 export const selectAuth = (state: RootState) => state.auth;
 export const authReducer = authSlice.reducer;
